Remove target version from source prechecking
The following removes the target version from source prechecking.
The source check should __never__ know about the version it's moving
to. The source doesn't care. The target cares. Removing this logic
is the first step of cleaning all of this up. Inreality we should
send a big set of information for the target controller to validate.
Nothing should be checked on the source.
