+++ upgrades/upgradevalidation/migrate.go
@@ -4,8 +4,6 @@
 package upgradevalidation
 
 import (
-	"github.com/juju/version/v2"
-
 	environscloudspec "github.com/juju/juju/environs/cloudspec"
 )
 
@@ -13,19 +11,13 @@ import (
 // controller for model migration.
 // Note: the target version can never be lower than the current version.
 func ValidatorsForModelMigrationSource(
-	targetVersion version.Number, cloudspec environscloudspec.CloudSpec,
+	cloudspec environscloudspec.CloudSpec,
 ) []Validator {
-	validators := []Validator{
-		getCheckTargetVersionForModel(targetVersion, MigrateToAllowed),
+	return []Validator{
 		getCheckUpgradeSeriesLockForModel(false),
 		checkNoWinMachinesForModel,
 		checkForDeprecatedUbuntuSeriesForModel,
 		getCheckForLXDVersion(cloudspec),
+		checkForCharmStoreCharms,
 	}
-	// If the target version is 3.1 or greater, we need to check for charm
-	// store charms and prevent them from being migrated.
-	if targetVersion.Major > 3 || (targetVersion.Major == 3 && targetVersion.Minor >= 1) {
-		validators = append(validators, checkForCharmStoreCharms)
-	}
-	return validators
 }
