+++ upgrades/upgradevalidation/migrate_test.go
@@ -8,7 +8,6 @@ import (
 	"github.com/juju/errors"
 	jujutesting "github.com/juju/testing"
 	jc "github.com/juju/testing/checkers"
-	"github.com/juju/version/v2"
 	gc "gopkg.in/check.v1"
 
 	environscloudspec "github.com/juju/juju/environs/cloudspec"
@@ -67,12 +66,11 @@ type migrateSuite struct {
 }
 
 func (s *migrateSuite) TestValidatorsForModelMigrationSourceJuju3(c *gc.C) {
-	ctrl, cloudSpec := s.setupJuju3Target(c)
+	ctrl, cloudSpec := s.setupMocks(c)
 	defer ctrl.Finish()
 
 	modelTag := coretesting.ModelTag
-	targetVersion := version.MustParse("3.0.0")
-	validators := upgradevalidation.ValidatorsForModelMigrationSource(targetVersion, cloudSpec)
+	validators := upgradevalidation.ValidatorsForModelMigrationSource(cloudSpec)
 
 	checker := upgradevalidation.NewModelUpgradeCheck(modelTag.Id(), s.statePool, s.st, s.model, validators...)
 	blockers, err := checker.Validate()
@@ -81,15 +79,11 @@ func (s *migrateSuite) TestValidatorsForModelMigrationSourceJuju3(c *gc.C) {
 }
 
 func (s *migrateSuite) TestValidatorsForModelMigrationSourceJuju31(c *gc.C) {
-	ctrl, cloudSpec := s.setupJuju3Target(c)
+	ctrl, cloudSpec := s.setupMocks(c)
 	defer ctrl.Finish()
 
-	// - check no charm store charms
-	s.st.EXPECT().AllCharmURLs().Return([]*string{}, errors.NotFoundf("charm urls"))
-
 	modelTag := coretesting.ModelTag
-	targetVersion := version.MustParse("3.1.0")
-	validators := upgradevalidation.ValidatorsForModelMigrationSource(targetVersion, cloudSpec)
+	validators := upgradevalidation.ValidatorsForModelMigrationSource(cloudSpec)
 
 	checker := upgradevalidation.NewModelUpgradeCheck(modelTag.Id(), s.statePool, s.st, s.model, validators...)
 	blockers, err := checker.Validate()
@@ -105,7 +99,7 @@ func (s *migrateSuite) initializeMocks(c *gc.C) *gomock.Controller {
 	return ctrl
 }
 
-func (s *migrateSuite) setupJuju3Target(c *gc.C) (*gomock.Controller, environscloudspec.CloudSpec) {
+func (s *migrateSuite) setupMocks(c *gc.C) (*gomock.Controller, environscloudspec.CloudSpec) {
 	ctrl := s.initializeMocks(c)
 	server := mocks.NewMockServer(ctrl)
 	serverFactory := mocks.NewMockServerFactory(ctrl)
@@ -119,13 +113,13 @@ func (s *migrateSuite) setupJuju3Target(c *gc.C) (*gomock.Controller, environscl
 			return serverFactory
 		},
 	)
-	// - check agent version;
-	s.model.EXPECT().AgentVersion().Return(version.MustParse("2.9.36"), nil)
 	// - check no upgrade series in process.
 	s.st.EXPECT().HasUpgradeSeriesLocks().Return(false, nil)
 	// - check if the model has win machines;
 	s.st.EXPECT().MachineCountForBase(makeBases("windows", winVersions)).Return(nil, nil)
 	s.st.EXPECT().MachineCountForBase(makeBases("ubuntu", ubuntuVersions)).Return(nil, nil)
+	// - check no charm store charms
+	s.st.EXPECT().AllCharmURLs().Return([]*string{}, errors.NotFoundf("charm urls"))
 
 	return ctrl, cloudSpec.CloudSpec
 }
